+++ library/src/main/java/com/getstream/sdk/chat/utils/StringUtility.java
@@ -4,6 +4,9 @@ package com.getstream.sdk.chat.utils;
 import androidx.annotation.Nullable;
 import android.text.TextUtils;
 
+import com.getstream.sdk.chat.rest.Message;
+import com.getstream.sdk.chat.rest.User;
+
 public class StringUtility {
 
     public static String stringFromNumbers(int... numbers) {
@@ -29,5 +32,21 @@ public class StringUtility {
                 "[\u203C\u2049]\uFE0F?|[\u25AA\u25AB\u25B6\u25C0\u25FB-\u25FE]\uFE0F?|" +
                 "[\u00A9\u00AE]\uFE0F?|[\u2122\u2139]\uFE0F?|\uD83C\uDC04\uFE0F?|\uD83C\uDCCF\uFE0F?|" +
                 "[\u231A\u231B\u2328\u23CF\u23E9-\u23F3\u23F8-\u23FA]\uFE0F?)+");
+    }
+
+    public static String getDeletedOrMentionedText(Message message) {
+        if (message == null) return null;
+        String text = message.getText();
+        if (message.getDeletedAt() != null) {
+            text = "_" + Constant.MESSAGE_DELETED + "_";
+            return text;
+        }
+        if (message.getMentionedUsers() != null && !message.getMentionedUsers().isEmpty()) {
+            for (User mentionedUser : message.getMentionedUsers()) {
+                String userName = mentionedUser.getName();
+                text = text.replace("@" + userName, "**" + "@" + userName + "**");
+            }
+        }
+        return text;
     }
 }
