+++ library/src/main/java/com/getstream/sdk/chat/utils/Utils.java
@@ -262,20 +262,4 @@ public class Utils {
         }
         return mentionedUserIDs;
     }
-
-    public static String getDeletedOrMentionedText(Message message) {
-        if (message == null) return null;
-        String text = message.getText();
-        if (message.getDeletedAt() != null) {
-            text = "_" + message.getText() + "_";
-            return text;
-        }
-        if (message.getMentionedUsers() != null && !message.getMentionedUsers().isEmpty()) {
-            for (User mentionedUser : message.getMentionedUsers()) {
-                String userName = mentionedUser.getName();
-                text = text.replace("@" + userName, "**" + "@" + userName + "**");
-            }
-        }
-        return text;
-    }
 }
